# internal/handler/user_handler.go
package handler

import (
	"encoding/json"
	"log/slog"
	"net/http"

	"github.com/valeroman/devboard/internal/domain"
	"github.com/valeroman/devboard/internal/usecase"
	"github.com/valeroman/devboard/internal/validator"
)

// UserHandler traduce request HTTP a las llamadas del UserUseCase
type UserHandler struct {
	usecase   *usecase.UserUseCase
	validator *validator.Validator
	logger    *slog.Logger
}

// NewUserHandler (contructor) crea un nuevo UserHandler con las dependencies proporcionadas.
func NewUserHandler(uc *usecase.UserUseCase, v *validator.Validator, logger *slog.Logger) *UserHandler {
	return &UserHandler{
		usecase:   uc,
		validator: v,
		logger:    logger,
	}
}

type createUserRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Name     string `json:"name" validate:"required,min=2,max=100"`
	Password string `json:"password" validate:"required,min=8"`
}

type userResponse struct {
	ID    string `json:"id"`
	Email string `json:"email"`
	Name  string `json:"name"`
}

// Create user endpoint maneja la creación de un nuevo usuario.
//
// @Summary Crear usuario
// @Description Registra un nuevo usuario en el sistema
// @Tags users
// @Accept json
// @Produce json
// @Param user body createUserRequest true "Datos del usuario a crear"
// @Success 201 {object} userResponse
// @Failure 400 {object} ProblemDetails
// @Router /users [post]
func (h *UserHandler) Create(w http.ResponseWriter, r *http.Request) {
	var req createUserRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		RespondError(w, r, h.logger, domain.ErrInvalidInput)
		return
	}

	if errs := h.validator.Validate(req); errs != nil {
		if err := RespondValidationError(w, errs); err != nil {
			h.logger.ErrorContext(r.Context(), "error al escribir respuesta de validación", "error", err, "method", r.Method, "path", r.URL.Path)
		}
		return
	}

	// TEMPORAL !!!!!
	passwordHash := "TEMPORAL_" + req.Password

	user, err := h.usecase.CreateUser(r.Context(), req.Email, req.Name, passwordHash)

	if err != nil {
		RespondError(w, r, h.logger, err)
	}

	if err := RespondJSON(w, http.StatusCreated, userResponse{
		ID:    user.ID,
		Email: user.Email,
		Name:  user.Name,
	}); err != nil {
		h.logger.ErrorContext(r.Context(), "error al escribir respuesta JSON al crear usuario", "error", err, "method", r.Method, "path", r.URL.Path)
	}
}

// Get user endpoint maneja la obtención de un usuario por su ID.
//
// @Summary Obtener usuario
// @Description Obtiene un usuario por su ID
// @Tags users
// @Produce json
// @Param id path string true "ID del usuario"
// @Success 200 {object} userResponse
// @Failure 404 {object} ProblemDetails
// @Router /users/{id} [get]
func (h *UserHandler) Get(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")

	user, err := h.usecase.GetUser(r.Context(), id)

	if err != nil {
		RespondError(w, r, h.logger, err)
		return
	}

	if err := RespondJSON(w, http.StatusOK, userResponse{
		ID:    user.ID,
		Email: user.Email,
		Name:  user.Name,
	}); err != nil {
		h.logger.ErrorContext(
			r.Context(),
			"error al escribir respuesta JSON al obtener usuario",
			"error",
			err,
			"method",
			r.Method,
			"path",
			r.URL.Path,
		)
	}
}